backend/pyproject.toml
[project]
name = "solarpulse-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.10.1",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra"
asyncio_mode = "auto"
markers = [
  "unit: tests puros sin I/O (backend/tests/unit)",
  "integration: cruzan capas via MockTransport/filesystem, sin red externa (backend/tests/integration)",
  "tinku: integracion HTTP real al middleware Tinku (solo con --tinku)",
  "offline: lectura de snapshots locales sin red",
  "http_contract: contrato de codigos HTTP aguas arriba sin red",
]
